requirements.txt
google-adk
fastapi
uvicorn[standard]
jinja2
